+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 0.2
-- FEATURE: Added /cgive pastebin [id]
+- FEATURE: Added '/cgive pastebin [id]' command
+- FEATURE: Added '/cobalt reload' command
 
 # Version 0.1
 Initial version